drivers: wifi: Fix IDLE thread stack address corruption
In case VIF addition fails we return out of bound array value (max
value) and this was assigned first and boundary checked later.

This ended up overwriting the IDLE thread stack address to "2" (address
in flash) leading to an obscure crash with no proper call stack.

Fix the assignment after the boundary check.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>